project-drf/requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# requirement

Django==2.2.6
djangorestframework==3.8.2
django-rest-swagger==2.2.0
babel==2.11.0
xlwt==1.3.0
